requirements.txt
awesome-slugify
beautifulsoup4
Pillow
pypandoc
python-dateutil
requests
toml
